refactor(docker): rename rhel-smoke.py to playwright-smoke.py for cross-platform use
